update ledrail.scad
Add a cone for the screw heads and use actual screw parameters for
length and width.

Remove wago / 5V IN compartment and reduce complexity.

Widen the cable duct on the inside for more space of the three cables
between the esp and the strip.

Integrate the back side-plate and only use one removal for accessing the
esp and led strip.

Introduce a tolerance variable to take into account that the acryl glass
is not always exactly as thick as advertised.